refactor(products): implement ProductPhotosUploader component
- Create a new ProductPhotosUploader component for uploading product photos
- Replace image upload logic in CreateProductPage and EditProductPage with ProductPhotosUploader
- Remove redundant code and simplify a photo upload process
