+++ src/app/stores/[storeId]/products/[productId]/edit/page.tsx
@@ -1,12 +1,11 @@
 'use client';
 
-import { Button, Form, Input, ScrollShadow, Textarea } from '@heroui/react';
+import { Button, Form, Input, Textarea } from '@heroui/react';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { hapticFeedback } from '@telegram-apps/sdk';
-import Image from 'next/image';
 import { useParams, useRouter } from 'next/navigation';
-import React, { useState } from 'react';
-import { useFieldArray, useForm, useFormContext } from 'react-hook-form';
+import React from 'react';
+import { useFieldArray, useForm } from 'react-hook-form';
 import toast from 'react-hot-toast';
 
 import AppLayout from '@/libs/common/components/AppLayout';
@@ -14,12 +13,9 @@ import { PageHeader } from '@/libs/common/components/page-header';
 import { ProductAttributeFields } from '@/libs/products/components/product-attributes-field';
 import { ProductTypeSelector } from '@/libs/products/components/product-type-selector';
 import { ProductVariantFields } from '@/libs/products/components/product-variants-field';
-import {
-  useProductDetails,
-  useUpdateProductMutation,
-  useUploadProductPhotosMutation,
-} from '@/libs/products/hooks';
+import { useProductDetails, useUpdateProductMutation } from '@/libs/products/hooks';
 import { UpdateProductFormData, updateProductSchema } from '@/libs/products/schemas';
+import { ProductPhotosUploader } from '@/libs/products/components/product-photos-uploader';
 
 export default function EditProductPage() {
   const { storeId, productId } = useParams<{ storeId: string; productId: string }>();
@@ -48,10 +44,7 @@ export default function EditProductPage() {
     },
   });
 
-  const setValue = useFormContext().setValue;
-
   const productType = watch('productType');
-  const [previewUrls, setPreviewUrls] = useState<string[]>([]);
 
   const {
     fields: attributeFields,
@@ -65,23 +58,8 @@ export default function EditProductPage() {
     remove: removeVariant,
   } = useFieldArray({ control, name: 'variants' });
 
-  const { mutateAsync: mutateAsyncProductPhotos, isPending } = useUploadProductPhotosMutation();
-
   const { mutateAsync } = useUpdateProductMutation(storeIdNum, productIdNum);
   const router = useRouter();
-
-  const handleOnChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const files = Array.from(e.target.files || []);
-    if (!files.length) return;
-
-    try {
-      const result = await mutateAsyncProductPhotos(files);
-      setPreviewUrls(result.imageUrls);
-      setValue('imageUrls', result.imageUrls);
-    } catch (err) {
-      toast.error('Failed to upload images');
-    }
-  };
 
   const onSubmit = async (data: UpdateProductFormData) => {
     try {
@@ -99,6 +77,8 @@ export default function EditProductPage() {
     <AppLayout>
       <Form onSubmit={handleSubmit(onSubmit)}>
         <PageHeader title="Edit Product" />
+
+        <ProductPhotosUploader />
 
         <Input
           label="Product Name"
@@ -120,22 +100,6 @@ export default function EditProductPage() {
           {...register('description')}
           placeholder="Write a short product description..."
         />
-
-        <Input
-          label="Image URL"
-          {...register('imageUrls')}
-          onChange={handleOnChange}
-          isInvalid={!!errors.imageUrls}
-          errorMessage={errors.imageUrls?.message}
-        />
-
-        {previewUrls.length && (
-          <ScrollShadow orientation="horizontal" className="flex gap-x-2">
-            {previewUrls.map((url) => (
-              <Image key={url} src={url} width={100} height={100} className="rounded" alt="" />
-            ))}
-          </ScrollShadow>
-        )}
 
         <ProductTypeSelector name="productType" control={control} errors={errors} />
